Merge #1258
1258: NO-TICKET: limit deploy size to 1MiB r=Fraser999 a=Fraser999

This PR limits the deploy size, both when accepting one in the node, and when creating one in the client.

The node was previously rejecting deploys greater than the permitted maximum block size.  It now rejects deploys with a serialized size of > 1MiB.  The serialized size is that provided by the `ToBytes` trait for the `Deploy` type.

The client also now checks that any deploy created is within this limit and avoids sending such an invalid deploy to the node.

Co-authored-by: Fraser Hutchison <fraser@casperlabs.io>